refactor: MovieItem 컴포넌트 ts로 migration
