+++ src/components/MovieList/MovieItem/style.css
